+++ assets/javascripts/kitten/components/buttons/button.js
@@ -29,7 +29,7 @@ export class Button extends Component {
     )
 
     // Adds keyboard accessibility to `<a>`
-    const tabindex = (tag == "a" && !this.props.href) ? 0 : null
+    const tabindex = (tag == 'a' && !this.props.href) ? 0 : null
 
     const Tag = tag