+++ assets/javascripts/kitten/components/form/text-input.js
@@ -11,14 +11,16 @@ export class TextInput extends Component {
   }
 
   render() {
-    const { className,
-            tag,
-            valid,
-            error,
-            disabled,
-            tiny,
-            digits,
-            ...others } = this.props
+    const {
+      className,
+      tag,
+      valid,
+      error,
+      disabled,
+      tiny,
+      digits,
+      ...others
+    } = this.props
 
     const textInputClassName = classNames(
       'k-TextInput',