+++ spec/dummy/client/entries/app-kitten.js
@@ -66,6 +66,7 @@ import {
 
 // Footer
 import { KarlFooterLendo } from 'kitten/karl/footer/footer-lendo'
+import { KarlFooterKisskiss } from 'kitten/karl/footer/footer-kisskiss'
 
 // Form
 import { Checkbox } from 'kitten/components/form/checkbox'
@@ -354,6 +355,7 @@ ReactOnRails.register({
 
   // Footer
   KarlFooterLendo,
+  KarlFooterKisskiss,
 
   // Form
   Checkbox,