+++ CHANGELOG.md
@@ -4,6 +4,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Fix: Fix `margin-bottom` on `LinkList` with `margin` prop.
+- Feature: Add `k-u-color-background1` utility for `Text` component.
+- Feature: Add `focus()` to `TextInput`.
+
 ## [19.3.1] - 2018-01-31
 
 Fix: